// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/clinical-new/shared/model/add-vitals-object.model.ts
export class AddVitalsObjectModel {
    PatientId: number = 0;
    PatientVisitId: number = 0;
    VitalsId: number = 0;
    VitalsValue: string = "";
    Unit: string = "";
    OtherVariable: string = "";
    Remarks: string = "";
    IsActive: boolean = false;
    VitalsCode: string = "";
    VitalsType: string = "";
    VitalsName: string = "";
    DisplayOrder: number = 0;
    VitalsGroup: string = "";
    InputType: string = "";
}